ChatPDF/requirements.txt
langchain
langchain-community
langchain-openai
openai
pypdf
streamlit
python-dotenv
faiss-cpu

